pyproject.toml
[project]
name = "google-ads-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29",
    "google-ads>=24.0.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "boto3>=1.34",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "alembic>=1.13",
    "openai>=1.30.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff"]

[tool.setuptools.packages.find]
include = ["backend*", "mcp_server*", "agents*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
